docker/vertexai-proxy/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
httpx>=0.28.1
google-auth>=2.38.0
requests>=2.31.0
